+++ praktikumsplaner-frontend/frontend/src/components/assign/PraktikumsstelleCard.vue
@@ -92,8 +92,6 @@ import { findStudiengangColorByValue } from "@/types/Studiengang";
 import { findAusbildungsrichtungColorByValue } from "@/types/Ausbildungsrichtung";
 import { useWarnings } from "@/composables/warnings";
 import { useTextGenerator } from "@/composables/textGenerator";
-import { valueToNameStudiensemester } from "@/types/Studiensemester";
-import { valueToNameAusbildungsjahr } from "@/types/Ausbildungsjahr";
 
 const props = defineProps<{
     value: Praktikumsstelle;
@@ -104,50 +102,18 @@ const unassignDialogContent = ref<string>("");
 const unassignDialogTitle = ref<string>("Zuweisung aufheben?");
 const unassignConfirmDialog = ref<boolean>(false);
 const warningDialog = ref<boolean>(false);
-const stelleToAssignUnassign = ref<Praktikumsstelle>();
 const nwkToAssignUnassing = ref<Nwk>();
 const warningDialogTitle = ref<string>(
     "Warnung. Wollen sie wirklich fortfahren?"
 );
 const warningDialogText = ref<string>("");
 const assignedNwk = ref(props.value.assignedNwk);
+const generator = useTextGenerator();
+
+let stelleToAssignUnassign: Praktikumsstelle | undefined;
 
 function getCardText(stelle: Praktikumsstelle): string {
-    let cardText = "";
-
-    if (stelle.studiengang) {
-        cardText += "Studiengang: " + stelle.studiengang + "\n";
-        if (stelle.studiensemester) {
-            cardText += "Semester: ";
-            stelle.studiensemester.sort();
-            for (let i = 0; i < stelle.studiensemester.length - 1; i++) {
-                cardText +=
-                    valueToNameStudiensemester(stelle.studiensemester[i]) +
-                    ", ";
-            }
-            cardText +=
-                valueToNameStudiensemester(
-                    stelle.studiensemester[stelle.studiensemester.length - 1]
-                ) + "\n";
-        }
-    } else if (stelle.ausbildungsrichtung) {
-        cardText += "Ausbildungsrichtung: " + stelle.ausbildungsrichtung + "\n";
-        if (stelle.ausbildungsjahr) {
-            cardText += "Ausbildungsjahr: ";
-            stelle.ausbildungsjahr.sort();
-            for (let i = 0; i < stelle.ausbildungsjahr.length; i++) {
-                cardText +=
-                    valueToNameAusbildungsjahr(stelle.ausbildungsjahr[i]) +
-                    ", ";
-            }
-            cardText +=
-                valueToNameAusbildungsjahr(
-                    stelle.ausbildungsjahr[stelle.ausbildungsjahr.length - 1]
-                ) + "\n";
-        }
-    }
-
-    return cardText;
+    return generator.getPraktikumsstellenCardText(stelle);
 }
 
 function getCardDetailText(stelle: Praktikumsstelle): string {
@@ -182,114 +148,11 @@ function drop(event: DragEvent, stelle: Praktikumsstelle) {
         return;
     }
 
-    // Check if Studiums or Ausbildungspraktikumsstelle
-    if (
-        stelle.ausbildungsrichtung == undefined &&
-        nwkToAssignUnassing.value.studiengang == undefined
-    ) {
-        warningDialogText.value +=
-            "Wollen sie wirklich " +
-            nwkToAssignUnassing.value.vorname +
-            " " +
-            nwkToAssignUnassing.value.nachname +
-            " auf eine Studiumspraktikumsstelle setzen, obwohl er/sie Auszubildende/r ist?\n";
-    }
-
-    if (
-        stelle.studiengang == undefined &&
-        nwkToAssignUnassing.value.ausbildungsrichtung == undefined
-    ) {
-        warningDialogText.value +=
-            "Wollen sie wirklich " +
-            nwkToAssignUnassing.value.vorname +
-            " " +
-            nwkToAssignUnassing.value.nachname +
-            " auf eine Ausbildungspraktikumsstelle setzen, obwohl er/sie Student*in ist?\n";
-    }
-
-    // Check if studiengang is the same
-    if (
-        stelle.studiengang &&
-        nwkToAssignUnassing.value?.studiengang != undefined &&
-        stelle.studiengang != nwkToAssignUnassing.value.studiengang
-    ) {
-        warningDialogText.value +=
-            "Wollen sie wirklich eine/n " +
-            nwkToAssignUnassing.value.studiengang +
-            " Student*in auf eine " +
-            stelle.studiengang +
-            " Stelle setzen?\n";
-    }
-
-    // Check if requested Nwk is the same
-    if (
-        stelle.namentlicheAnforderung &&
-        stelle.namentlicheAnforderung?.toUpperCase() !=
-            nwkToAssignUnassing.value.vorname.toUpperCase() +
-                " " +
-                nwkToAssignUnassing.value.nachname.toUpperCase()
-    ) {
-        warningDialogText.value +=
-            "Wollen sie wirklich " +
-            nwkToAssignUnassing.value.vorname +
-            " " +
-            nwkToAssignUnassing.value.nachname +
-            " auf diese Stelle setzen obwohl explizit " +
-            stelle.namentlicheAnforderung +
-            " angefordert wurde?\n";
-    }
-
-    // Check if Nwk is in the right semester
-    if (
-        stelle.studiengang != undefined &&
-        nwkToAssignUnassing.value?.studiengang != undefined &&
-        stelle.studiensemester
-    ) {
-        const expectedSemesters: number[] = [];
-        for (let i = 0; i < stelle.studiensemester.length; i++) {
-            expectedSemesters.push(+stelle.studiensemester[i].substring(8, 9));
-        }
-        const actualSemester = calculateSemester(nwkToAssignUnassing.value);
-        if (!expectedSemesters.includes(actualSemester)) {
-            warningDialogText.value +=
-                "Wollen sie wirklich eine/n Student*in im " +
-                actualSemester +
-                " Semester auf diese Stelle setzen, obwohl ein/e Student*in im ";
-            for (let i = 0; i < expectedSemesters.length; i++) {
-                warningDialogText.value += expectedSemesters[i] + ". Semester";
-                if (i < expectedSemesters.length - 1) {
-                    warningDialogText.value += ", ";
-                }
-            }
-            warningDialogText.value += " erwartet wird.\n";
-        }
-    }
-
-    // Check if Nwk is in the right Lehrjahr
-    if (
-        stelle.ausbildungsrichtung != undefined &&
-        nwkToAssignUnassing.value?.ausbildungsrichtung != undefined &&
-        stelle.ausbildungsjahr
-    ) {
-        const expectedLehrjahre: number[] = [];
-        for (let i = 0; i < stelle.ausbildungsjahr.length; i++) {
-            expectedLehrjahre.push(+stelle.ausbildungsjahr[i].substring(4, 5));
-        }
-        const actualLehrjahr = calculateLehrjahr(nwkToAssignUnassing.value);
-        if (!expectedLehrjahre.includes(actualLehrjahr)) {
-            warningDialogText.value +=
-                "Wollen sie wirklich eine/n Auszubildende/n im " +
-                actualLehrjahr +
-                ". Lehrjahr auf diese Stelle setzen, obwohl ein/e Auszubildende/r im ";
-            for (let i = 0; i < expectedLehrjahre.length; i++) {
-                warningDialogText.value += expectedLehrjahre[i] + ". Lehrjahr";
-                if (i < expectedLehrjahre.length - 1) {
-                    warningDialogText.value += ", ";
-                }
-            }
-            warningDialogText.value += " erwartet wird.\n";
-        }
-    }
+    let warnings = useWarnings().getBeforeAssignmentWarnings(
+        stelle,
+        nwkToAssignUnassing.value
+    );
+    warnings.forEach((w) => (warningDialogText.value += w.message + "\n"));
 
     stelleToAssignUnassign = stelle;
     if (warningDialogText.value == "") {
@@ -318,34 +181,6 @@ function assignNwk() {
 function resetWarningDialog() {
     warningDialogText.value = "";
     warningDialog.value = false;
-}
-
-function calculateSemester(nwk: Nwk) {
-    if (!nwk) return -1;
-    if (nwk && nwk.ausbildungsrichtung != undefined) return 0;
-    let semester: number;
-    const startYear: number = +nwk.jahrgang.substring(0, 2) + 2000;
-    const currentYear: number = new Date().getFullYear();
-    const difference = currentYear - startYear;
-    semester = difference * 2;
-    if (new Date().getMonth() > 8) semester += 1;
-    if (new Date().getMonth() < 3) semester -= 1;
-    return semester;
-}
-
-function calculateLehrjahr(nwk: Nwk) {
-    if (!nwk) return -1;
-    if (nwk.studiengang != undefined) return 0;
-    let lehrjahr: number;
-    const startYear: number = +nwk.jahrgang.substring(0, 2) + 2000;
-    const currentYear: number = new Date().getFullYear();
-    lehrjahr = currentYear - startYear;
-    if (new Date().getMonth() > 8) {
-        lehrjahr += 1;
-    } else {
-        lehrjahr -= 1;
-    }
-    return lehrjahr;
 }
 
 function unassignNwk() {
